core: Modernize C-style casts and apply architectural fixes

+++ src/core/model/time.cc
@@ -47,8 +47,7 @@ const int32_t UNIT_COEFF[Time::LAST] = {315360, 864, 36, 6, 1, 1, 1, 1, 1, 1};
 long double
 Scale(Time::Unit u)
 {
-    return UNIT_COEFF[static_cast<int>(u)] *
-           std::pow(10.0L, static_cast<long double>(UNIT_POWER[static_cast<int>(u)]));
+    return UNIT_COEFF[u] * std::pow(10.0L, UNIT_POWER[u]);
 }
 
 /**
@@ -219,17 +218,17 @@ Time::SetResolution(Unit unit, Resolution* resolution, const bool convert /* = t
 
     for (int i = 0; i < Time::LAST; i++)
     {
-        int shift = UNIT_POWER[i] - UNIT_POWER[static_cast<int>(unit)];
+        int shift = UNIT_POWER[i] - UNIT_POWER[unit];
         int quotient = 1;
-        if (UNIT_COEFF[i] > UNIT_COEFF[static_cast<int>(unit)])
+        if (UNIT_COEFF[i] > UNIT_COEFF[unit])
         {
-            quotient = UNIT_COEFF[i] / UNIT_COEFF[static_cast<int>(unit)];
-            NS_ASSERT(quotient * UNIT_COEFF[static_cast<int>(unit)] == UNIT_COEFF[i]);
+            quotient = UNIT_COEFF[i] / UNIT_COEFF[unit];
+            NS_ASSERT(quotient * UNIT_COEFF[unit] == UNIT_COEFF[i]);
         }
-        else if (UNIT_COEFF[i] < UNIT_COEFF[static_cast<int>(unit)])
+        else if (UNIT_COEFF[i] < UNIT_COEFF[unit])
         {
-            quotient = UNIT_COEFF[static_cast<int>(unit)] / UNIT_COEFF[i];
-            NS_ASSERT(quotient * UNIT_COEFF[i] == UNIT_COEFF[static_cast<int>(unit)]);
+            quotient = UNIT_COEFF[unit] / UNIT_COEFF[i];
+            NS_ASSERT(quotient * UNIT_COEFF[i] == UNIT_COEFF[unit]);
         }
         NS_LOG_DEBUG("SetResolution for unit " << +unit << " loop iteration " << i << " has shift "
                                                << shift << " has quotient " << quotient);
@@ -245,7 +244,7 @@ Time::SetResolution(Unit unit, Resolution* resolution, const bool convert /* = t
         }
         auto factor = static_cast<int64_t>(std::pow(10, std::fabs(shift)) * quotient);
         double realFactor = std::pow(10.0, static_cast<double>(shift)) *
-                            static_cast<double>(UNIT_COEFF[i]) / UNIT_COEFF[static_cast<int>(unit)];
+                            static_cast<double>(UNIT_COEFF[i]) / UNIT_COEFF[unit];
         NS_LOG_DEBUG("SetResolution factor " << factor << " real factor " << realFactor);
         info->factor = factor;
         // here we could equivalently check for realFactor == 1.0 but it's better
